Major (incompatible!) interface change.
The first argument of any handler is from now on the Kernel.
EvCtx was removed completely. To achieve what EvCtx provided,
the Kernel got context() API which provides the same,
and event more--overloaded context(back int) allows to explore
the full stack.
Because we don't know (there is no API to get it,
and likely won't be) what depth the stack is, context(int) form
returns a pointer, while context() - a reference.
Both context(0) and context(-1) do guarantee to return
non-NULL pointers.
0, 1, 2, ... means top-to-bottom direction, while
-1, -2, -3, ... the opposite bottom-to-top
(similarly as array indices in Perl).
context() is just a shorthand form of *context(0).
When a handler is called by the coexx,
kernel.context(-1) refers to what is at the bottom of
the call stack--context related to dispatched asynchronous
event which was taken out from the event loop,
that is, POST, ALARM, SELECT, SIGNAL, or IDLE (missed in enum).

As this change will force monkey changes in existing
apps, I let myself btw to rename DatIO to IO_Ctx,
and its field filedes to fileno, which both sound better, IMO.

Enjoy ;)