Intensive I/O activity could block alarms to expire.
The solution is to check expiration time of alarm events against actual time
regardless the select() returned 0 (explicit timeout) or positive value
(I/O ready).
It is necessary because in a face of big I/O activity, and small reads/writes,
the select() will still return value > 0 even in the case of the shortest
possible timeout, i.e. zero--the condition called a "polling".