Retry select() if reading notification pipe failed.
Failure with EINTR or EAGAIN is harmless, but if the pipe
is not read in _select_io call the following could happen:
if along with data in the pipe there are other descriptors ready,
they will cause pqueue events, thus _select_io to return.
on subsequent _select_io if other thread managed to "signal"
the pipe again before select+read call, the read will return
(assuming not interrupted again) 2 bytes, which is *by design*
never expected (I still assume 1 bytes is max. what can be
in the pipe at any given time).