List of Kernels in a Thread no longer rwlock-protected.
Before this change a remote thread could insert a moved kernel
to _list_kernel in a local (current) thread.
Because it could happen at any time, the operation was guarded
by rwlock. Therefore, any usage of _list_kernel, including
trivial tests, would have to be protected.
This would be neither efficient nor necessary. It might be even
problematic to treverse such a list in future API if such was added.

Because of all these reasons, it's much better to pass kernel
being moved from one thread to other at well defined points, i.e.
1) in source thread: remove from _list_kernel in _export_kernel_local_data()
   and next
2) in target thread: insert   to _list_kernel in _import_kernel_local_data()

Future implementations should be aware that there is a small window where
a kernel might not be present in any of _list_kernel(s), but this should
not be a problem because kernel->_thread is guaranteed to be always
valid and if changed, changed atomicaly (in _export_kernel_local_data
under rwlock)--this property is crucial since post'ing events must be
possible all the time, including a period when a kernel is in transient
state being moved to the other thread.