Fixed "oversignalled" communication pipe
The bug was with assumption that writing a byte to the internal comm
pipe (used to wake up the thread when other threads post events to it)
has immediate effect on the select(), and the successive read returns
that one byte (and not more). However, in the intensive I/O apps, even
just after a write to the pipe, other descriptors setup in the app might
already have the select prepared to return, so the observable effect of
the signalling the pipe will be deferred until next select call (I
believe not later than that). When the I/O traffic was really high, it
sometimes happened to read 2 bytes from the pipe (deferred + successive
current).