Rebuild sets for select() after timeout.
Cause of the change is similar as in the previous commit.

However it is not as evident but may results from my earlier observation
that select() on Linux can return 0 immediately (compared pre- and post-
timestamps) in case when given timeout was clearly > 0 (I have to
confirm it once again sometime).

If this happened, no timer event will be pqueue'd, thus inner loop
reiterates, and without rebuilding the sets for the select(), the strace
output as quoted above might be easily explained.

In the changed code the inner loop is no longer necessary, so it was
removed (add -b when git diff); Consistently, break statements were
changed to continue(s).