CoeStr - optimized string class for coexx
CoeStr is mostly used internally, but in external interface it is
visible as:
- input value (implicitly as conversion from const char*
and std::string, so the user may even not notice its existence)
- return value (a few rerely used functions like current state,
so the user may accept its limitted API (primary goal was performance,
not to double std::string)).

Current implementation relies on the information from /proc/self/maps to
determine whether a given c-string resides in initialized constant
memory region.

Because this may not be supported on particular operating system or not
work as expected (Cygwin?) the behavior is controlled by COEXX_COESTR
define. If not provided to a compiler explicitly, for Linux it defaults
to 3 (test/literal.cpp shows it works well), for other systems to 0.

COEXX_COESTR
    0 : fallback to std::string (safe but inefficient)
    1 : no runtime checks
    2 : runtime detection of string literals; assert/throw on failure
    3 : like 2 but conversion to std::string on failure (safe and efficient)