[incompatible] Extract Postback (handle) from Callback (pointer).
The whole change is planned as two phases just to ease updates in
existing applications. This commit extracts 'post' API from the Callback
class forming new class Postback, and at once making its objects the
handles, reference-counted objects which heap-based nature is internally
maintained by the library. In the next commit remaining Callback will be
converted to behave as the handle.

The change is incompatible and the motivation for it is as follow.

So far the callbacks were heap-based objects which clients were required
to delete them eventually. Obviously it was the opportunity for memory
leaks. Other downside was that, if passed to vparam(), wrapping in
owned_ptr<> template was tedious.

There are still heap objects like those returned by vparam() and
rparam() functions, but typically they are passed immediately as
arguments to various APIs of the library.

Therefore the conversion from user visible heap-based Callback*
(pointers) to real handles is definitely worthwile.

The rationale for the split of post & call APIs of the old Callback
class among new Postback and changed Callback classes is more subtle.

Old callbacks when created could have provided by vparam() so-called
"prefix" parameters. Such a prefix was cloned if a callback has been
post'ed, or could be modified if a callback has been call'ed.

Since "callbacks as pointers" were obviously owned by single sessions,
dual post/call interface could not be invoked at the same time, so it
never imposed any problem on the prefix.

Differently would be in "callbacks as handles" approach. If such
callbacks retained dual post/call interface, due to the ease of copying
the handles, a given callback might be quickly owned by multiple
sessions, each potentially under different thread. One can imagine that
one session could be in the midst of the call operation (thus modifying
the prefix), while the other session copying such the prefix (with
compromised data integrity) as part of the post on the callback.

Thus the split of post/call APIs along with new postbacks and callbacks
as handles will allow to keep data integrity of the prefix.

TODO:
- _Atomic<T> needs implementation not requiring GCC >= 4.1.2(?)
- Compiles on Solaris SPARC with GCC 4.1.1, but links with error:
  undefined reference to `__sync_sub_and_fetch_4'.
  Check if there is any -march=... that solves the problem,
  similarly as on 64-bit Linux -march=i486 option does
  when you have to compile/link with -m32.