ContextSwitch -- switching the context to another session.
ContextSwitch object *must* be auto (on the stack) variable.

If a given session as the parameter to ContextSwitch constructor
has ever been started, the 'kernel' method will return non-NULL pointer,
even if that session is about to stop, or already stopped.

If the switch is successful (kernel() returning non-NULL),
effective kernel()->context() is characterized by:
- type() being EventContext::CALL, and
- state() name ".switch".

Obviously, the session reference must represent valid Session object.
It's easy to assure if the _given_ session is e.g. an ancestor
of the _current_ session.