Fix/improve assertions that check clobbering of internal msgpipe(s).
Bug in the assertion (coe-thread--d4t.cpp, line 608) revealed
when _msgpipe_rfd was greater than max. user descriptor scheduled to read
(or temporarily there was no descriptor to read at all),
and 'lvec' buffer current length (lvec_max) was too short,
so the assertion, checking a bit for _msgpipe_rfd, could get the bit SET
because it was in out-of-buffer space.

Fix by rearranging the code, keeping all 'lvec' buffers (read/write/error)
of the same size, including place for 'msgpipe_wfd', even if it's never put in,
but this simplifies further checks, and more, there are even more checks now,
because a user code might clobber descriptor of read pipe for writing
or descriptor of write pipe for reading!